requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
channels[daphne]==4.0.0
Django==4.1.7
django-crispy-forms==1.14.0
django-js-lib-tabulator==4.9.3.0
google-api-core==2.11.0
googleapis-common-protos==1.58.0
google-api-python-client==2.90.0
google-auth==2.16.2
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-cloud-pubsub
google-crc32c==1.5.0
google-resumable-media==2.4.1
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
install==1.3.5
mssql-django==1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
db-dtypes==1.1.1
Pillow==9.4.0
pip==23.0.1
protobuf==4.22.1
proto-plus==1.22.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
setuptools==49.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
gunicorn==20.1.0
psycopg2-binary==2.9.5
django-storages[google]==1.12.3
django-environ==0.8.